ocp/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.1
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.5.0
soupsieve==2.4
sqlparse==0.4.3
